% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_sequence.R
\name{aa_sequence}
\alias{aa_sequence}
\title{Write aa sequence from codons}
\usage{
aa_sequence(codons)
}
\arguments{
\item{codons}{list of codons and corresponding letter}
}
\value{
A character vector of amino acids
}
\description{
Write aa sequence from codons
}
\examples{
codons <- c("UUU","UUC")
aa_sequence(codons)
}
