% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colplot_counts_aa_in_proteinseq.R
\name{colplot_number_aa_in_protseq}
\alias{colplot_number_aa_in_protseq}
\title{counts number of different aminosacids in a protein sequence and plot sit in a colplot}
\usage{
colplot_number_aa_in_protseq(protein_sequence)
}
\arguments{
\item{protein_sequence}{a string of aa}
}
\value{
plot of number of aminoacids in a column plot
}
\description{
counts number of different aminosacids in a protein sequence and plot sit in a colplot
}
