% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Codon Extractor}
\usage{
extract_codons(seq_string, start = 1)
}
\arguments{
\item{seq_string}{The dna or rna sequence for which you want to extract codons}

\item{start}{the position from which the reading frame should begin}
}
\value{
returns a vector of codons for the given sequence and reading frame
}
\description{
Codon Extractor
}
