% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_T_to_U.R
\name{from_T_to_U}
\alias{from_T_to_U}
\title{from T to U}
\usage{
from_T_to_U(getting_T)
}
\arguments{
\item{getting_T}{This is where we get T nucleotide in DNA sequence.}
}
\value{
From here we give U nucleotide in RNA sequence.
}
\description{
from T to U
}
\examples{
from_T_to_U("ATCGGCATTCGAATCG")

}
