% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{Generate DNA sequence}
\usage{
generate_dna(length)
}
\arguments{
\item{length}{integer, corresponds to the desired length of the sequence}
}
\value{
string representing DNA sequence
}
\description{
Generates a sequence of given length of bases A,T,G,C
}
\examples{
generate_dna(10)
generate_dna(50)

}
