% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_sequence}
\alias{generate_sequence}
\title{generate_sequence}
\usage{
generate_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{the length of the sequence you want to generate}
}
\value{
Random DNA sequence
}
\description{
generate_sequence
}
\examples{
sequence_length <- 10
dna_sequence <- generate_sequence(sequence_length)
}
