% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_to_codon.R
\name{RNA_to_codon}
\alias{RNA_to_codon}
\title{Converting an RNA sequence to codons}
\usage{
RNA_to_codon(RNA_seq, start = 1)
}
\arguments{
\item{RNA_seq, }{a sequence of RNA}

\item{start, }{the position in the sequence you want to start converting into codons. Default is 1}
}
\value{
A sequence of codons derived from the input RNA
}
\description{
Converting an RNA sequence to codons
}
