% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_to_protein.R
\name{RNA_to_protein}
\alias{RNA_to_protein}
\title{RNA to protein}
\usage{
RNA_to_protein(codons_list)
}
\arguments{
\item{codons_list}{The list of codons}
}
\value{
The protein sequence from given list of codons
}
\description{
RNA to protein
}
\examples{
x <- c("UUA", "AUG", "GUA")
RNA_to_protein(x)
}
