% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription}
\usage{
transcription(DNA)
}
\arguments{
\item{DNA}{A DNA sequence}
}
\value{
A RNA sequence
}
\description{
Transcription
}
\examples{
DNA_sequence <- "CTGGTGGTTTAGGCGTACAATGTCCTGAAGAATATTTAAGAAAAAAGCACCCCTCGTCGCCTAGAATTACCTACCGCGGTCGACCATACCTTCGATTATC"
transcription(DNA_sequence)
}
